MAIN CHANGE: new services layer, important mods in file structure
