+++ controllers/UserController.js
@@ -1,20 +1,21 @@
-import { UserModel } from '../models/UserModel.js'
+import { UserServices } from '../services/UserServices.js'
 import {AppError} from '../utils/appError.js'
 import { validatePartialSchema, validateSchema } from '../utils/validateSchema.js'
-import { userSchema } from '../validations/userValidations.js'
+import { userSchema } from '../validations/userSchema.js'
+import { successResponse } from '../utils/response.js'
 import bcrypt from 'bcrypt'
 
 export class UserController{
     static async getAllUsers(req,res){
-        const users  = await UserModel.getAllUsers()
-        return res.json(users)
+        const users  = await UserServices.getAllUsers()
+        successResponse(res,users)
     }
 
-    static async getUserById(req,res,next){
+    static async getUserById(req,res){
         const {id} = req.params
-        const user = await UserModel.getUserById(id)
+        const user = await UserServices.getUserById(id)
         if (!user) throw new AppError('User not found', 404)
-        res.json(user)
+        successResponse(res,user)
     }
 
     static async createUser(req,res){
@@ -26,9 +27,8 @@ export class UserController{
         const {password} = result.data
         const passwordHashed = await bcrypt.hash(password,10)
 
-        const newUser = await UserModel.createUser({...result.data, password: passwordHashed})
-        res.status(201).json(newUser)
-
+        const newUser = await UserServices.createUser({...result.data, password: passwordHashed})
+        successResponse(res,newUser,201)
     }
     static async updateUser(req,res){
         const result = validatePartialSchema(userSchema, req.body)
@@ -37,14 +37,14 @@ export class UserController{
         }
 
         const {id} = req.params
-        const userUpdated = await UserModel.updateUser(id, result.data)
+        const userUpdated = await UserServices.updateUser(id, result.data)
         if (!userUpdated) throw new AppError('User not found', 404)
-        res.json(userUpdated) 
+        successResponse(res,userUpdated) 
     }
     static async deleteUser(req,res ){
         const {id} = req.params
-        const userDeleted = await UserModel.deleteUser(id)
+        const userDeleted = await UserServices.deleteUser(id)
         if (!userDeleted) throw new AppError('User not found', 404)
-        res.json(userDeleted)
+        successResponse(res,userDeleted)
     }
 }
