+++ utils/response.js
@@ -1,0 +1,6 @@
+export const successResponse = (res, data, statusCode = 200) => {
+  res.status(statusCode).json({
+    success: true,
+    data,
+  })
+}
